refactor: 修复 clippy 警告 + 重构心跳数据结构
主要修改：
- 添加 HeartbeatData 结构体封装心跳参数（解决 too_many_arguments）
- 修复 never_loop 警告（for -> if let）
- 添加 Default 实现到多个结构体
- 修复 no_effect_replace 警告
- 自动应用 clippy --fix 建议（67 处）

+++ src/security/injection.rs
@@ -79,7 +79,7 @@ impl PromptInjectionDetector {
         
         for (pattern, threat_type, severity, description) in ignore_patterns {
             if let Ok(re) = Regex::new(pattern) {
-                patterns.push((re, threat_type, severity as u8, description.to_string()));
+                patterns.push((re, threat_type, severity, description.to_string()));
             }
         }
         
@@ -95,7 +95,7 @@ impl PromptInjectionDetector {
         
         for (pattern, threat_type, severity, description) in roleplay_patterns {
             if let Ok(re) = Regex::new(pattern) {
-                patterns.push((re, threat_type, severity as u8, description.to_string()));
+                patterns.push((re, threat_type, severity, description.to_string()));
             }
         }
